Move core management to its own component
