+++ src/components/Main.ui.js
@@ -2,29 +2,9 @@
 
 import { type DiskletFolder, makeReactNativeFolder } from 'disklet'
 import { ModalManager } from 'edge-components'
-import type { EdgeContext, EdgeCorePluginFactory, EdgeCurrencyPlugin } from 'edge-core-js'
-import { eosCurrencyPluginFactory, rippleCurrencyPluginFactory, stellarCurrencyPluginFactory } from 'edge-currency-accountbased'
-import {
-  bitcoinCurrencyPluginFactory,
-  bitcoincashCurrencyPluginFactory,
-  bitcoingoldCurrencyPluginFactory,
-  bitcoinsvCurrencyPluginFactory,
-  dashCurrencyPluginFactory,
-  digibyteCurrencyPluginFactory,
-  feathercoinCurrencyPluginFactory,
-  groestlcoinCurrencyPluginFactory,
-  litecoinCurrencyPluginFactory,
-  qtumCurrencyPluginFactory,
-  smartcashCurrencyPluginFactory,
-  ufoCurrencyPluginFactory,
-  vertcoinCurrencyPluginFactory,
-  zcoinCurrencyPluginFactory
-} from 'edge-currency-bitcoin'
-import { ethereumCurrencyPluginFactory } from 'edge-currency-ethereum'
-import { moneroCurrencyPluginFactory } from 'edge-currency-monero'
-import { coinbasePlugin, coincapPlugin, currencyconverterapiPlugin, shapeshiftPlugin } from 'edge-exchange-plugins'
+import type { EdgeContext, EdgeCurrencyPlugin } from 'edge-core-js'
 import React, { Component } from 'react'
-import { Image, Keyboard, Linking, StatusBar, TouchableWithoutFeedback, View, YellowBox } from 'react-native'
+import { Alert, Image, Keyboard, Linking, StatusBar, TouchableWithoutFeedback, View, YellowBox } from 'react-native'
 import DeviceInfo from 'react-native-device-info'
 import Locale from 'react-native-locale'
 import { MenuProvider } from 'react-native-popup-menu'
@@ -83,10 +63,7 @@ import * as Constants from '../constants/indexConstants'
 import { scale } from '../lib/scaling.js'
 import { setIntlLocale } from '../locales/intl'
 import s, { selectLocale } from '../locales/strings.js'
-import EdgeAccountCallbackManager from '../modules/Core/Account/EdgeAccountCallbackManager.js'
 import * as CONTEXT_API from '../modules/Core/Context/api'
-import EdgeContextCallbackManager from '../modules/Core/Context/EdgeContextCallbackManager.js'
-import EdgeWalletsCallbackManager from '../modules/Core/Wallets/EdgeWalletsCallbackManager.js'
 import DeepLinkingManager from '../modules/DeepLinkingManager.js'
 import PermissionsManager, { type Permission, PermissionStrings } from '../modules/PermissionsManager.js'
 import AutoLogout from '../modules/UI/components/AutoLogout/AutoLogoutConnector'
@@ -107,41 +84,11 @@ import TransactionAlert from '../modules/UI/components/TransactionAlert/Transact
 import { PluginBuySell, PluginSpend, PluginView, renderPluginBackButton } from '../modules/UI/scenes/Plugins/index'
 import { HwBackButtonHandler } from '../modules/UI/scenes/WalletList/components/HwBackButtonHandler/index'
 import { styles } from '../styles/MainStyle.js'
-import { makeCoreContext } from '../util/makeContext.js'
+import { EdgeCoreManager } from './core/EdgeCoreManager.js'
 import { CreateWalletName } from './scenes/CreateWalletNameScene.js'
 import { CryptoExchangeQuoteProcessingScreenComponent } from './scenes/CryptoExchangeQuoteProcessingScene.js'
 import { OnBoardingComponent } from './scenes/OnBoardingScene.js'
 import { TermsOfServiceComponent } from './scenes/TermsOfServiceScene.js'
-
-const pluginFactories: Array<EdgeCorePluginFactory> = [
-  // Exchanges:
-  coinbasePlugin,
-  shapeshiftPlugin,
-  coincapPlugin,
-  currencyconverterapiPlugin,
-  // Currencies:
-  bitcoincashCurrencyPluginFactory,
-  bitcoinCurrencyPluginFactory,
-  ethereumCurrencyPluginFactory,
-  eosCurrencyPluginFactory,
-  stellarCurrencyPluginFactory,
-  rippleCurrencyPluginFactory,
-  moneroCurrencyPluginFactory,
-  dashCurrencyPluginFactory,
-  litecoinCurrencyPluginFactory,
-  bitcoinsvCurrencyPluginFactory,
-  // eboostCurrencyPluginFactory,
-  // dogecoinCurrencyPluginFactory,
-  qtumCurrencyPluginFactory,
-  digibyteCurrencyPluginFactory,
-  zcoinCurrencyPluginFactory,
-  bitcoingoldCurrencyPluginFactory,
-  vertcoinCurrencyPluginFactory,
-  feathercoinCurrencyPluginFactory,
-  smartcashCurrencyPluginFactory,
-  groestlcoinCurrencyPluginFactory,
-  ufoCurrencyPluginFactory
-]
 
 const localeInfo = Locale.constants() // should likely be moved to login system and inserted into Redux
 
@@ -214,9 +161,6 @@ type Props = {
   deepLinkPending: boolean,
   checkAndShowGetCryptoModal: () => void
 }
-type State = {
-  context: ?EdgeContext
-}
 
 async function queryUtilServer (context: EdgeContext, folder: DiskletFolder, usernames: Array<string>) {
   let jsonObj: null | Object = null
@@ -251,7 +195,7 @@ async function queryUtilServer (context: EdgeContext, folder: DiskletFolder, use
   }
 }
 
-export default class Main extends Component<Props, State> {
+export default class Main extends Component<Props> {
   keyboardDidShowListener: any
   keyboardDidHideListener: any
 
@@ -259,9 +203,6 @@ export default class Main extends Component<Props, State> {
     super(props)
     slowlog(this, /.*/, global.slowlogOptions)
 
-    this.state = {
-      context: undefined
-    }
     if (ENV.HIDE_IS_MOUNTED) {
       YellowBox.ignoreWarnings([
         'Warning: isMounted(...) is deprecated',
@@ -285,24 +226,7 @@ export default class Main extends Component<Props, State> {
     const id = DeviceInfo.getUniqueID()
     global.firebase && global.firebase.analytics().setUserId(id)
     global.firebase && global.firebase.analytics().logEvent(`Start_App`)
-    makeCoreContext(pluginFactories).then(context => {
-      const folder = makeReactNativeFolder()
 
-      // Put the context into Redux:
-      this.props.addContext(context, folder)
-
-      CONTEXT_API.listUsernames(context).then(usernames => {
-        this.props.addUsernames(usernames)
-        queryUtilServer(context, folder, usernames)
-      })
-      setIntlLocale(localeInfo)
-      selectLocale(DeviceInfo.getDeviceLocale())
-      SplashScreen.close({
-        animationType: SplashScreen.animationType.fade,
-        duration: 850,
-        delay: 500
-      })
-    })
     Linking.getInitialURL()
       .then(url => {
         if (url) {
@@ -312,6 +236,34 @@ export default class Main extends Component<Props, State> {
       })
       .catch(e => console.log(e))
     Linking.addEventListener('url', this.handleOpenURL)
+  }
+
+  onCoreLoad = (context: EdgeContext) => {
+    const folder = makeReactNativeFolder()
+
+    // Put the context into Redux:
+    this.props.addContext(context, folder)
+
+    CONTEXT_API.listUsernames(context).then(usernames => {
+      this.props.addUsernames(usernames)
+      queryUtilServer(context, folder, usernames)
+    })
+    setIntlLocale(localeInfo)
+    selectLocale(DeviceInfo.getDeviceLocale())
+    SplashScreen.close({
+      animationType: SplashScreen.animationType.fade,
+      duration: 850,
+      delay: 500
+    })
+  }
+
+  onCoreError = (error: any) => {
+    SplashScreen.close({
+      animationType: SplashScreen.animationType.fade,
+      duration: 850,
+      delay: 500
+    })
+    Alert.alert('Edge core failed to load', String(error))
   }
 
   handleOpenURL = (event: Object) => {
@@ -778,9 +730,7 @@ export default class Main extends Component<Props, State> {
         <ModalManager />
         <PermissionsManager />
 
-        <EdgeContextCallbackManager />
-        <EdgeAccountCallbackManager />
-        <EdgeWalletsCallbackManager />
+        <EdgeCoreManager onLoad={this.onCoreLoad} onError={this.onCoreError} />
 
         <DeepLinkingManager />
       </MenuProvider>
